refactor(firmware): split ModbusSlave into per-entity modules
Break the monolithic ModbusSlave.{cpp,h} into 10 cohesive modbus_*
translation units, each owning one concern and its own build gate.
Behavior-preserving; validated by arduino-cli builds (RTU single-serial,
debug-only/TCP, dual-serial).

- modbus_config.h   build gates (defines.h + MB_SERIAL_ACTIVE/DEBUG_* derived)
- modbus_types.h    enums, MBinfo, frame-size/status constants
- modbus_frame.*    shared seam (mb_frame/mb_frame_len/modbus, exceptionResponse)
- modbus_crc.*      CRC-16 + tables (defined once; fixes latent per-TU flash dup)
- modbus_registers.* register store + operation FCs (#ifdef MODBUS_ENABLED)
- modbus_debug.*    debugger FCs 0x41-0x48 (home for future licensing FCs)
- modbus_pdu.*      process_mbpacket + mb_pdu_request_len/mb_pdu_skips_crc
- modbus_serial.*   RTU/debugger serial transport (single + dual-serial)
- modbus_tcp.*      Modbus TCP transport (Ethernet/WiFi/ETH)
- ModbusSlave.*     umbrella header + mbtask() facade (Baremetal.ino unchanged)

Key decoupling: the serial transport no longer knows the function-code set.
It asks modbus_pdu for per-FC frame shape (mb_pdu_request_len) and CRC policy
(mb_pdu_skips_crc), so adding a function code touches only modbus_pdu +
its handler, never the transports.
