+++ resources/sources/Baremetal/modbus_debug.h
@@ -1,0 +1,29 @@
+/*
+modbus_debug.h - OpenPLC always-on debugger function codes (0x41-0x48)
+Copyright (C) 2022 OpenPLC - Thiago Alves
+
+The debugger PDU handlers, dispatched from process_mbpacket. Kept ungated: the
+dispatch in modbus_pdu references them unconditionally (the always-on debugger is
+present on every baremetal build). This is the growth home for future custom FCs
+(e.g. the 0x49-0x4C licensing set).
+*/
+
+#ifndef MODBUS_DEBUG_H
+#define MODBUS_DEBUG_H
+
+#include "modbus_frame.h"
+
+// Phase 4 debugger entrypoints. Signatures changed from MatIEC-era
+// (flat u16 index) to the (array_idx: u8, elem_idx: u16) addressing model.
+void debugInfo(void);
+void debugSetTrace(uint8_t arr, uint16_t elem, uint8_t flag,
+                   uint16_t len, void *value);
+void debugGetTrace(uint8_t arr, uint16_t startidx, uint16_t endidx);
+void debugGetTraceList(uint16_t numIndexes, uint8_t *indexArray);
+void debugGetMd5(void *endianness);
+// Always-on debugger extras — served even without full Modbus (DEBUGGER_ENABLED).
+void debugGetStatus(void);
+void debugGetVersion(void);
+void debugGetBoardId(void);
+
+#endif
